Ading tests and fixing grammar
